Allow manual action, fix deps, refs and copy pasta
